Corrections de bugs au niveau des enchantements, ajout des comportements associés aux enchantements (fire aspect, fire arrows donnent des cooked items et looting augmentent les chances de drops).
git-svn-id: https://xp-dev.com/svn/WoIR/trunk@131 8b22c5f6-99f7-4f02-aeed-e951476ab6e7

+++ WorldOfIceRealm.RareDrops/src/ca/qc/icerealm/bukkit/plugins/raredrops/RareDropsFactory.java
@@ -10,9 +10,9 @@ import ca.qc.icerealm.bukkit.plugins.raredrops.enchantment.EnchantmentsOdds;
 public class RareDropsFactory 
 {
 	public final Logger logger = Logger.getLogger(("Minecraft"));
-	private int oddsMultiplier = 1;
+	private double oddsMultiplier = 1;
 		
-	public int getOddsMultiplier() 
+	public double getOddsMultiplier() 
 	{
 		return oddsMultiplier;
 	}
@@ -27,7 +27,7 @@ public class RareDropsFactory
 		
 	}
 	
-	public RareDropsFactory(int _oddsMultiplier)
+	public RareDropsFactory(double _oddsMultiplier)
 	{
 		oddsMultiplier = _oddsMultiplier;
 	}
